Refactor config handling, add env utils, and improve SDK initialization
